relocate audio for rendering tests
